requirements.txt
streamlit
google.generativeai
pillow
datetime
